Restructuring folders.
