+++ cibyl/outputs/cli/ci/system/impls/zuul/colored.py
@@ -18,11 +18,12 @@ import logging
 from overrides import overrides
 
 from cibyl.cli.query import QueryType
-from cibyl.outputs.cli.ci.systems.base.colored import ColoredBaseSystemPrinter
-from cibyl.outputs.cli.ci.systems.common.builds import (get_duration_section,
-                                                        get_status_section)
-from cibyl.outputs.cli.ci.systems.common.jobs import (get_plugin_section,
-                                                      has_plugin_section)
+from cibyl.outputs.cli.ci.system.common.builds import (get_duration_section,
+                                                       get_status_section)
+from cibyl.outputs.cli.ci.system.common.jobs import (get_plugin_section,
+                                                     has_plugin_section)
+from cibyl.outputs.cli.ci.system.impls.base.colored import \
+    ColoredBaseSystemPrinter
 from cibyl.outputs.cli.printer import ColoredPrinter
 from cibyl.utils.strings import IndentedTextBuilder
